ap_rputs() silently truncated size_t to int, resulting in 130+ warnings when building httpd and a rich set of modules on Windows.
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@1529449 13f79535-47bb-0310-9956-ffa450edef68

+++ include/http_protocol.h
@@ -411,7 +411,7 @@ AP_DECLARE(int) ap_rwrite(const void *buf, int nbyte, request_rec *r);
  */
 static APR_INLINE int ap_rputs(const char *str, request_rec *r)
 {
-    return ap_rwrite(str, strlen(str), r);
+    return ap_rwrite(str, (int)strlen(str), r);
 }
 
 /**
